fix: unpack int64/uint64 proto fields via i64, not i32
All four integer-constructor branches (optional scalar, repeated scalar,
plain int32/uint32, plain int64/uint64) were calling
<i32 as UnpackValue>::unpack_value and casting with `as _`.  This
rejects or silently truncates any value outside i32 range before the
cast, so fields like Digest.size_bytes > 2 GiB and any uint32 > 2^31-1
would fail or produce wrong results.

Switch to <i64 as UnpackValue>::unpack_value for all non-enum integer
fields; starlark-rust implements UnpackValue for i64 and handles the
full 64-bit range.  Enum fields stay on i32 since proto enum values are
always small.

+++ crates/starbuf-derive/src/lib.rs
@@ -536,7 +536,7 @@ fn try_message(input: TokenStream) -> Result<TokenStream, Error> {
                 }
             } else if attrs.optional && (attrs.int32 || attrs.uint32 || attrs.int64 || attrs.uint64) {
                 quote! {
-                    result.#fident = Some(<i32 as ::starlark::values::UnpackValue>::unpack_value(value)?
+                    result.#fident = Some(<i64 as ::starlark::values::UnpackValue>::unpack_value(value)?
                         .ok_or_else(|| ::anyhow::anyhow!("expected int"))? as _);
                 }
             } else if attrs.optional && attrs.bool {
@@ -574,7 +574,7 @@ fn try_message(input: TokenStream) -> Result<TokenStream, Error> {
                     let list = ::starlark::values::list::ListRef::from_value(value)
                         .ok_or_else(|| ::anyhow::anyhow!("field '{}' expects a list", #fident_str))?;
                     for item in list.iter() {
-                        result.#fident.push(<i32 as ::starlark::values::UnpackValue>::unpack_value(item)?
+                        result.#fident.push(<i64 as ::starlark::values::UnpackValue>::unpack_value(item)?
                             .ok_or_else(|| ::anyhow::anyhow!("expected int"))? as _);
                     }
                 }
@@ -600,12 +600,12 @@ fn try_message(input: TokenStream) -> Result<TokenStream, Error> {
                 }
             } else if attrs.int32 || attrs.uint32 {
                 quote! {
-                    result.#fident = <i32 as ::starlark::values::UnpackValue>::unpack_value(value)?
+                    result.#fident = <i64 as ::starlark::values::UnpackValue>::unpack_value(value)?
                         .ok_or_else(|| ::anyhow::anyhow!("expected int"))? as _;
                 }
             } else if attrs.int64 || attrs.uint64 {
                 quote! {
-                    result.#fident = <i32 as ::starlark::values::UnpackValue>::unpack_value(value)?
+                    result.#fident = <i64 as ::starlark::values::UnpackValue>::unpack_value(value)?
                         .ok_or_else(|| ::anyhow::anyhow!("expected int"))? as _;
                 }
             } else if attrs.enumeration.is_some() {
